Update AVCGoToGoal.java
Added a scaled down version of the figure 8 to see what the robot's motion did. Brooks, we never tested the robot doing a figure eight so you can't say we broke it. We changed the constants of the robot because the wheel diameter had gotten squished, and therefore smaller, we also had tweaked the wheel base and measured it from center of the wheel to center of the wheel. Adjusted the tolerance to be smaller.
*There are a few key points that'll bring the robot around the course*. The robot does a lot of curlies and whoop-die-dos before reaching the point, so we minimized the amount of points. The robot knows where it is, but it doesn't turn well. It makes an arc to get from point to point.